Added truckloads of comments

+++ src/main/java/com/justbuyit/eventprocessor/subscription/NotifySubscriptionEventProcessor.java
@@ -1,21 +1,20 @@
 package com.justbuyit.eventprocessor.subscription;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import com.justbuyit.auth.ConnectionSigner;
 import com.justbuyit.dao.CompanyDAO;
 import com.justbuyit.entity.Company;
 import com.justbuyit.eventprocessor.EventProcessor;
+import com.justbuyit.exception.AccountNotFoundException;
 import com.justbuyit.exception.JustBuyItException;
 import com.justbuyit.model.event.subscription.NotifySubscriptionEvent;
 import com.justbuyit.model.event.subscription.NotifySubscriptionEvent.NotifySubscriptionPayload;
 import com.justbuyit.model.result.Result;
 
+/**
+ * Event processor for subsciption notify events.
+ */
 public class NotifySubscriptionEventProcessor extends EventProcessor<NotifySubscriptionEvent> {
 
-    private final static Logger LOG = LoggerFactory.getLogger(NotifySubscriptionEventProcessor.class);
-    
     private CompanyDAO companyDAO;
 
     public NotifySubscriptionEventProcessor(ConnectionSigner connectionSigner, CompanyDAO companyDAO) {
@@ -25,19 +24,28 @@ public class NotifySubscriptionEventProcessor extends EventProcessor<NotifySubsc
 
     @Override
     protected Result processEvent(NotifySubscriptionEvent event) throws JustBuyItException {
-        LOG.debug("Processing event [{}]", event);
-
         String companyId = event.getPayload().getAccount().getAccountIdentifier();
         String noticeType = event.getPayload().getNotice().getType();
         if (noticeType.equals("CLOSED")) {
 
-            // delete company/subscription
+            // fetch company
             Company company = companyDAO.findById(companyId);
+            if (company == null) {
+                throw new AccountNotFoundException(String.format("Could not find account [%s]", companyId));
+            }
+            
+            // delete company/subscription
             companyDAO.delete(company);
+            
         } else if (!noticeType.equals("UPCOMING_INVOICE")) {
             
-            // update subscription status
+            // fetch company
             Company company = companyDAO.findById(companyId);
+            if (company == null) {
+                throw new AccountNotFoundException(String.format("Could not find account [%s]", companyId));
+            }
+
+            // update subscription status
             company.setSubscriptionStatus(event.getPayload().getAccount().getStatus());
             companyDAO.update(company);
         }
